Store the various admins inside `Market`
Store the various admins, such as `fee_collector`, `close_authority`,
etc. inside the `Market` struct.

+++ programs/openbook-v2/src/pod_option.rs
@@ -11,7 +11,7 @@ use std::convert::From;
 #[derive(Copy, Clone, Debug)]
 #[repr(C)]
 pub struct PodOption<T: Pod> {
-    flag: u8,
+    flag: u64,
     value: T,
 }
